// web/src/korrel8r/client/types.gen.ts
export type Class = string;

/**
 * Constrains the objects that will be included in search results.
 */
export type Constraint = {
    /**
     * Ignore objects with timestamps before this start time.
     */
    start?: string;
    /**
     * Ignore objects with timestamps after this end time.
     */
    end?: string;
    /**
     * Limit total number of objects per query.
     */
    limit?: number;
    /**
     * DEPRECATED store calls are cancelled with the request.
     */
    timeout?: Duration;
};

/**
 * The duration string is a sequence of decimal numbers, each with optional fraction and a unit suffix. Valid time units are: ns, us (or µs), ms, s, m, h.
 *
 */
export type Duration = string;

/**
 * Domain configuration information.
 */
export type Domain = {
    /**
     * Name of the domain.
     */
    name: string;
    /**
     * Brief description of the domain.
     */
    description?: string;
    /**
     * Stores configured for the domain.
     */
    stores?: Array<Store>;
};

/**
 * Directed edge in the result graph, from Start to Goal classes.
 */
export type Edge = {
    /**
     * Class name of the start node.
     */
    start: Class;
    /**
     * Class name of the goal node.
     */
    goal: Class;
    /**
     * Set of rules followed along this edge.
     */
    rules?: Array<Rule>;
};
// ...
export type Graph = {
    /**
     * List of graph edges.
     */
    edges?: Array<Edge>;
    /**
     * List of graph nodes.
     */
    nodes?: Array<Node>;
};
// ...
export type Node = {
    /**
     * Full class name
     */
    class: string;
    /**
     * Queries yielding results for this class.
     */
    queries?: Array<QueryCount>;
    /**
     * Number of results for this class, after de-duplication.
     */
    count?: number;
    /**
     * Serialized result contents, may be large.
     */
    result?: Array<Object>;
};

/**
 * Query with number of results.
 */
export type QueryCount = {
    /**
     * Number of results, omitted if the query was not executed.
     */
    count?: number;
    /**
     * Query for correlation data.
     */
    query: Query;
};

/**
 * Rule is a correlation rule with a list of queries and results counts found during navigation.
 */
export type Rule = {
    /**
     * Name is an optional descriptive name.
     */
    name: string;
    /**
     * Queries generated while following this rule.
     */
    queries?: Array<QueryCount>;
};

/**
 * Data object serialized as JSON.
 */
export type Object = {
    [key: string]: unknown;
};